+++ buildSrc/src/main/kotlin/CommonModulePlugin.kt
@@ -1,19 +1,11 @@
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.JavaVersion
-import org.gradle.api.plugins.ExtensionAware
-import org.gradle.kotlin.dsl.apply
-import org.gradle.kotlin.dsl.configure
+import org.gradle.api.artifacts.dsl.DependencyHandler
 
 class CommonModulePlugin: Plugin<Project> {
-//    private val Project.android: BaseExtension
-//        get() = extensions.findByName("android") as? BaseExtension
-//            ?: error("Not an Android module $name")
-
     override fun apply(project: Project) {
         with(project) {
             applyPlugins()
-//            androidConf()
             dependenciesConf()
         }
     }
@@ -23,38 +15,36 @@ class CommonModulePlugin: Plugin<Project> {
             apply("com.android.library")
             apply("kotlin-android")
             apply("com.vanniktech.maven.publish")
+            apply("shot")
+            apply("org.jetbrains.dokka")
         }
     }
 
-//    private fun Project.androidConf() {
-//        android.run {
-//            lintOptions.isAbortOnError = false
-//
-//            compileSdkVersion(30)
-//
-//            buildFeatures.compose = true
-//
-//            compileOptions {
-//                sourceCompatibility = JavaVersion.VERSION_1_8
-//                targetCompatibility = JavaVersion.VERSION_1_8
-//            }
-//
-//            composeOptions {
-//                kotlinCompilerExtensionVersion = Dependencies.AndroidX.Compose.version
-//            }
-//        }
-//    }
-
     private fun Project.dependenciesConf() {
         dependencies.apply {
-            add("implementation", Dependencies.Kotlin.stdlib)
-            add("implementation", Dependencies.AndroidX.coreKtx)
-            add("implementation", Dependencies.AndroidX.appcompat)
-            add("implementation", Dependencies.material)
+            implementation(Dependencies.Kotlin.stdlib)
+            implementation(Dependencies.AndroidX.coreKtx)
+            implementation(Dependencies.AndroidX.appcompat)
+            implementation(Dependencies.material)
 
-            add("implementation", Dependencies.AndroidX.Compose.ui)
-            add("implementation", Dependencies.AndroidX.Compose.material)
-            add("implementation", Dependencies.AndroidX.Compose.materialIconsExtended)
+            implementation(Dependencies.AndroidX.Compose.ui)
+            implementation(Dependencies.AndroidX.Compose.material)
+            implementation(Dependencies.AndroidX.Compose.materialIconsExtended)
+
+            androidTestImplementation(Dependencies.AndroidX.Compose.activity)
+            androidTestImplementation(Dependencies.AndroidX.Compose.testing)
+            androidTestImplementation(Dependencies.AndroidX.Testing.core)
+            androidTestImplementation(Dependencies.AndroidX.Testing.rules)
+            androidTestImplementation(Dependencies.AndroidX.Testing.runner)
+            add("androidTestImplementation", project(":test-utils"))
         }
+    }
+
+    private fun DependencyHandler.implementation(dependency: String) {
+        add("implementation", dependency)
+    }
+
+    private fun DependencyHandler.androidTestImplementation(dependency: String) {
+        add("androidTestImplementation", dependency)
     }
 }